+++ pytg/interfaces/cli_socket/__init__.py
@@ -4,5 +4,4 @@ __author__ = 'luckydonald'
 import logging
 logger = logging.getLogger(__name__)
 
-
 from . import receiver